Portfolio/requirements.txt
Django == 3.2.3
Pillow == 8.4.0
djangorestframework == 3.12.4
django-cors-headers == 3.7.0
black == 22.1.0